+++ src/Serializer/JsonApiSerializer.php
@@ -11,29 +11,32 @@
 
 namespace League\Fractal\Serializer;
 
-use RuntimeException;
-
 class JsonApiSerializer extends ArraySerializer
 {
     /**
-     * Serialize the top level data.
+     * Serialize an item resource
      *
      * @param string $resourceKey
      * @param array $data
      *
      * @return array
      */
-    public function serializeData($resourceKey, array $data)
+    public function collection($resourceKey, array $data)
     {
-        if (! $resourceKey) {
-            throw new RuntimeException('The $resourceKey parameter must be provided when using '.__CLASS__);
-        }
+        return array($resourceKey ?: 'data' => $data);
+    }
 
-        if (count($data) === count($data, COUNT_RECURSIVE)) {
-            $data = array($data);
-        }
-
-        return array($resourceKey => $data);
+    /**
+     * Serialize an item resource
+     *
+     * @param string $resourceKey
+     * @param array $data
+     *
+     * @return array
+     */
+    public function item($resourceKey, array $data)
+    {
+        return array($resourceKey ?: 'data' => array($data));
     }
 
     /**
@@ -41,9 +44,10 @@ class JsonApiSerializer extends ArraySerializer
      * 
      * @param  string  $resourceKey
      * @param  array  $data
+     *
      * @return array
      */
-    public function serializeIncludedData($resourceKey, array $data)
+    public function includedData($resourceKey, array $data)
     {
         $serializedData = array();
 
@@ -58,7 +62,7 @@ class JsonApiSerializer extends ArraySerializer
 
     /**
      * Indicates if includes should be side-loaded.
-     * 
+     *
      * @return bool
      */
     public function sideloadIncludes()