+++ test/ScopeTest.php
@@ -5,6 +5,8 @@ use League\Fractal\Resource\Collection;
 use League\Fractal\Resource\Item;
 use League\Fractal\Manager;
 use League\Fractal\Scope;
+use League\Fractal\Serializer\ArraySerializer;
+use League\Fractal\Test\Stub\Transformer\DefaultIncludeBookTransformer;
 use Mockery;
 
 class ScopeTest extends \PHPUnit_Framework_TestCase
@@ -160,12 +162,12 @@ class ScopeTest extends \PHPUnit_Framework_TestCase
 
     public function testToArrayWithSideloadedIncludes()
     {
-        $serializer = Mockery::mock('League\Fractal\Serializer\ArraySerializer[sideloadIncludes,serializeData,serializeIncludedData]');
+        $serializer = Mockery::mock('League\Fractal\Serializer\ArraySerializer')->makePartial();
         $serializer->shouldReceive('sideloadIncludes')->andReturn(true);
-        $serializer->shouldReceive('serializeData')->andReturnUsing(function ($key, $data) {
+        $serializer->shouldReceive('item')->andReturnUsing(function ($key, $data) {
             return array('data' => $data);
         });
-        $serializer->shouldReceive('serializeIncludedData')->andReturnUsing(function ($key, $data) {
+        $serializer->shouldReceive('includedData')->andReturnUsing(function ($key, $data) {
             return array('sideloaded' => array_pop($data));
         });
 
@@ -216,6 +218,7 @@ class ScopeTest extends \PHPUnit_Framework_TestCase
         $transformer = Mockery::mock('League\Fractal\TransformerAbstract');
         $transformer->shouldReceive('transform')->once()->andReturn($this->simpleItem);
         $transformer->shouldReceive('getAvailableIncludes')->once()->andReturn(array());
+        $transformer->shouldReceive('getDefaultIncludes')->once()->andReturn(array());
 
         $resource = new Item($this->simpleItem, $transformer);
         $scope = $manager->createData($resource);
@@ -229,6 +232,7 @@ class ScopeTest extends \PHPUnit_Framework_TestCase
         $transformer = Mockery::mock('League\Fractal\TransformerAbstract');
         $transformer->shouldReceive('transform')->once()->andReturn(array('foo' => 'bar'));
         $transformer->shouldReceive('getAvailableIncludes')->once()->andReturn(array());
+        $transformer->shouldReceive('getDefaultIncludes')->once()->andReturn(array());
 
         $resource = new Collection(array(array('foo' => 'bar')), $transformer);
         $scope = $manager->createData($resource);
@@ -239,7 +243,7 @@ class ScopeTest extends \PHPUnit_Framework_TestCase
     /**
      * @covers League\Fractal\Scope::executeResourceTransformers
      * @expectedException InvalidArgumentException
-     * @expectedExceptionMessage Argument $resource should be an instance of Resource\Item or Resource\Collection
+     * @expectedExceptionMessage Argument $resource should be an instance of League\Fractal\Resource\Item or League\Fractal\Resource\Collection
      */
     public function testCreateDataWithClassFuckKnows()
     {
@@ -341,6 +345,27 @@ class ScopeTest extends \PHPUnit_Framework_TestCase
         );
 
         $this->assertEquals($expectedOutput, $rootScope->toArray());
+    }
+
+    public function testDefaultIncludeSuccess()
+    {
+        $manager = new Manager();
+        $manager->setSerializer(new ArraySerializer());
+
+        // Send this stub junk, it has a specific format anyhow
+        $resource = new Item(array(), new DefaultIncludeBookTransformer());
+
+        // Try without metadata
+        $scope = new Scope($manager, $resource);
+
+        $expected = array(
+            'a' => 'b',
+            'author' => array(
+                'c' => 'd',
+            ),
+        );
+
+        $this->assertEquals($expected, $scope->toArray());
     }
 
     public function tearDown()