+++ CHANGELOG.md
@@ -1,8 +1,23 @@
-## 0.9.0 (2014-??-??)
+## 0.10.0 (2014-??-??)
 
 Features:
 
   - Added support for 4.2.x's more de-coupled pagination component
+
+## 0.8.3 (2014-06-14)
+
+Features:
+
+  - Default Includes no longer need to be in Available Includes. [Issue #58]
+
+[Issue #58]: https://github.com/thephpleague/fractal/issues/58
+
+## 0.8.2 (2014-06-09)
+
+Bug:
+
+  - A `null` value for `Manager::parseIncludes()` could have weird results
+>>>>>>> d6cd8b8afdf7f0deaccfbddc21b1a43954c8fbfc
 
 ## 0.8.1 (2014-06-05)